gevent_zeromq / __init__.py: naming update.
Hi and thanks for developing this nice library.
While packaging it for Debian I've noticed that almost everywhere in the code and docs the library is called gevent-zeromq or gevent_zeromq, apart from here and in the .tar.gz tarballs.
In order to prevent users confusion, would you consider changing the tarball names to gevent-zeromq? Thank you!